+++ okhttp-tests/src/test/java/com/squareup/okhttp/RequestTest.java
@@ -76,6 +76,6 @@ public final class RequestTest {
   private String bodyToHex(Request.Body body) throws IOException {
     OkBuffer buffer = new OkBuffer();
     body.writeTo(buffer);
-    return buffer.readByteString((int) buffer.size()).hex();
+    return buffer.readByteString(buffer.size()).hex();
   }
 }
